test/vhost: Fix vhost migration tc1 test
From qemu 6.1 The use of qemu-img create that depends on a backing
file now requires that an explicit backing format be provided.
This is for safety: if QEMU probes a different format than what you
thought, the data presented to the guest will be corrupt.

Fixes: #2466

Signed-off-by: Jaroslaw Chachulski <jaroslawx.chachulski@intel.com>
Change-Id: I56ae7c71097edb387b14c8a042408cd6c50af4ee
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/12268
Reviewed-by: Pawel Piatek <pawelx.piatek@intel.com>
Reviewed-by: Kamil Godzwon <kamilx.godzwon@intel.com>
Reviewed-by: Changpeng Liu <changpeng.liu@intel.com>
Reviewed-by: Karol Latecki <karol.latecki@intel.com>
Reviewed-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>
Community-CI: Broadcom CI <spdk-ci.pdl@broadcom.com>
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>