doc/doxyfile: remove timestamps generation
Doc files should be updated only if the file
content (actual documentation) will change.

Doxygen adds timestamps to the footer
of every doc file in spdk/doc directory.
Every day, when the date is changed,
all pages get a new date and it triggers
automatic commits push for all >500 doc pages
with a new date in the footer.

This change removes automatically generated
timestamps for each spdk/doc file footer.
In fact, these timestamps are not visible
directly on the page running in the browser anyway.

Signed-off-by: Kamil Godzwon <kamilx.godzwon@intel.com>
Change-Id: I6039a250613f109e255fd61a55ba9e3b197e7b10
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/12326
Reviewed-by: Konrad Sztyber <konrad.sztyber@intel.com>
Reviewed-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>