llvm_nvme_fuzz: raise SIGSEGV if user Ctrl-C
If an input causes a hang, the fuzzing thread won't
terminate itself, since it is waiting for all
outstanding commands to complete.  So raise a SIGSEGV
in the SPDK shutdown handler instead, which will
cause the fuzzer thread to exit as well as generating
an input file of the hung input.

Signed-off-by: Jim Harris <james.r.harris@intel.com>
Change-Id: I5753977740e27ca7827222b9e3cee1e939ef31a1
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/12407
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Community-CI: Broadcom CI <spdk-ci.pdl@broadcom.com>
Reviewed-by: Ben Walker <benjamin.walker@intel.com>
Reviewed-by: Changpeng Liu <changpeng.liu@intel.com>
Reviewed-by: <yifan.bian@intel.com>
Reviewed-by: Dong Yi <dongx.yi@intel.com>