test: process admin completions during stub sleep
The stub app runs during NVMe CI testing to minimize test time.
It runs as the primary process and all other NVMe tests
run as secondary processes and connect to the NVMe controller
via shared memory. The issue is when there is an Async Event
Notification, the secondary processes are not informed of the
event because the stub app does not process admin completions
and therefore they do not get added to other processes async
event list. This change adds a call to process admin
completions during the stub_sleep routine to cover this type
of AEN case. A configurable sleep time was also added to
allow the admin completion calls to occur more rapidly to help
minimize test time, if needed.

Signed-off-by: Curt Bruns <curt.e.bruns@gmail.com>
Change-Id: I72d5afc511c4409ec4a03cd969dbbc0d7dd4f256
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/12555
Community-CI: Broadcom CI <spdk-ci.pdl@broadcom.com>
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Shuhei Matsumoto <smatsumoto@nvidia.com>
Reviewed-by: Jim Harris <james.r.harris@intel.com>