test: optimize the print statements in aer test
The aer test print statements were overly complicated for multi
process mode, so a macro was created to simplify them. The macro
will prepend "[Child]" to the child process print statements to
make it easier to see which process is doing the printouts.

Signed-off-by: Curt Bruns <curt.e.bruns@gmail.com>
Change-Id: I0ecc5e2526b156fa7f98d44c745408ba922ebeff
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/12881
Community-CI: Broadcom CI <spdk-ci.pdl@broadcom.com>
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Paul Luse <paul.e.luse@intel.com>
Reviewed-by: Jim Harris <james.r.harris@intel.com>
Reviewed-by: Shuhei Matsumoto <smatsumoto@nvidia.com>