openstack: change transport type to tcp
Since we ended support for soft-roce, openstack
tests, have been failing. The cause is lack of
RDMA capable NICs in the setup. We do not need
to change the physical setup to add new NICs,
so we are changing the transport type to TCP.

Change-Id: I9dce0a4f1dc6dc51e22ba61a8c8a23f4178fde84
Signed-off-by: Krzysztof Karas <krzysztof.karas@intel.com>
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/13447
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Community-CI: Broadcom CI <spdk-ci.pdl@broadcom.com>
Reviewed-by: Konrad Sztyber <konrad.sztyber@intel.com>
Reviewed-by: John Kariuki <John.K.Kariuki@intel.com>
Reviewed-by: Jim Harris <james.r.harris@intel.com>